remove FSOBaseCLI, switch to use RocksDatabase in FSORepairTool, which wraps ManagedRocksDB
